refactor: 네트워크와 데이터베이스 작업 분리
- PlaceDBHelper: 로컬 데이터베이스에서 장소 정보를 가져오고 저장하는 기능 유지
