+++ app/src/main/java/campus/tech/kakao/map/PlaceApplication.kt
@@ -32,11 +32,8 @@ class PlaceApplication: Application() {
             val actNetwork: NetworkCapabilities =
                 connectivityManager.getNetworkCapabilities(network) ?: return false
 
-            return when {
-                actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> true
-                actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> true
-                else -> false
-            }
+            return actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) ||
+                    actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
         }
     }
 }
