+++ app/src/main/java/campus/tech/kakao/map/data/net/KakaoApiClient.kt
@@ -2,7 +2,10 @@ package campus.tech.kakao.map.data.net
 
 
 import campus.tech.kakao.map.BuildConfig
+import campus.tech.kakao.map.domain.model.Place
 import campus.tech.kakao.map.domain.model.ResultSearchKeyword
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -17,4 +20,21 @@ object KakaoApiClient {
             .build()
             .create(KakaoApi::class.java)
     }
+
+    suspend fun getPlaces(keyword: String): List<Place> =
+        withContext(Dispatchers.IO) {
+            val resultPlaces = mutableListOf<Place>()
+            for (page in 1..3) {
+                val response = api.getSearchKeyword(
+                    key = BuildConfig.KAKAO_REST_API_KEY,
+                    query = keyword,
+                    size = 15,
+                    page = page
+                )
+                if (response.isSuccessful) {
+                    response.body()?.documents?.let { resultPlaces.addAll(it) }
+                } else throw RuntimeException("통신 에러 발생")
+            }
+            resultPlaces
+        }
 }
